# packages/ai-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Service API
  description: |
    Multi-provider AI orchestration service with context optimization and intelligent project analysis.
    
    This service provides unified access to multiple AI providers (OpenAI, Anthropic, Google, Azure) with:
    - Intelligent context optimization
    - Project analysis and insights
    - Performance analytics and monitoring
    - Embedding generation
    - Provider management and health monitoring
  version: 1.0.0
  contact:
    name: CycleTime Team
    email: team@cycletime.dev
  license:
    name: MIT
servers:
  - url: http://localhost:8003
    description: Development server
  - url: https://ai-service-staging.cycletime.dev
    description: Staging server
  - url: https://ai-service.cycletime.dev
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/models:
    get:
      tags: [Model Management]
      summary: List available models
      description: Get list of available AI models across all providers
      operationId: listModels
      parameters:
        - name: provider
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ProviderId'
        - name: capability
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ModelCapability'
      responses:
        '200':
          description: List of available models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
  /api/v1/models/{modelId}:
    get:
      tags: [Model Management]
      summary: Get model details
      description: Get details of a specific AI model
      operationId: getModel
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDetails'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Chat Controller
...
  /api/v1/context/analyze:
    post:
      tags: [Context Optimization]
      summary: Analyze context
      description: Analyze context and optimize for AI consumption
      operationId: analyzeContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextAnalysisRequest'
      responses:
        '202':
          description: Context analysis started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextAnalysisResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/context/{contextId}/optimize:
    post:
      tags: [Context Optimization]
      summary: Optimize context
      description: Optimize context for specific model and use case
      operationId: optimizeContext
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextOptimizationRequest'
      responses:
        '200':
          description: Context optimization completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextOptimizationResponse'
        '404':
          description: Context not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/context/{contextId}/chunks:
    get:
      tags: [Context Optimization]
      summary: Get context chunks
      description: Get optimized context chunks for AI consumption
      operationId: getContextChunks
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: maxTokens
          in: query
          required: false
          schema:
            type: integer
            minimum: 100
            maximum: 200000
            default: 8000
        - name: priority
          in: query
          required: false
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0.5
      responses:
        '200':
          description: Context chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextChunksResponse'
        '404':
          description: Context not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Embedding Controller
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProviderId:
      type: string
      enum: [openai, anthropic, google, azure]
    ModelCapability:
      type: string
      enum: [text-generation, code-generation, embedding, vision, function-calling]
...
    Model:
      type: object
      required: [id, name, description, provider, capabilities, contextWindow, maxTokens, costPerInputToken, costPerOutputToken, status]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        provider:
          type: string
        capabilities:
          type: array
          items:
            type: string
        contextWindow:
          type: integer
        maxTokens:
          type: integer
        costPerInputToken:
          type: number
        costPerOutputToken:
          type: number
        status:
          type: string
    ModelsResponse:
      type: object
      required: [models, totalModels, lastUpdated]
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/Model'
        totalModels:
          type: integer
        lastUpdated:
          type: string
          format: date-time
    ModelDetails:
      type: object
      required: [id, name, description, provider, capabilities, contextWindow, maxTokens, costPerInputToken, costPerOutputToken, status, lastUpdated]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        provider:
          type: string
        capabilities:
          type: array
          items:
            type: string
        contextWindow:
          type: integer
        maxTokens:
          type: integer
        costPerInputToken:
          type: number
        costPerOutputToken:
          type: number
        status:
          type: string
        lastUpdated:
          type: string
          format: date-time

    # Chat schemas
...
    Document:
      type: object
      required: [id, content, title, type]
      properties:
        id:
          type: string
        content:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [markdown, text, code, json, yaml, typescript, javascript, python, java, go, rust]
        language:
          type: string
        metadata:
          type: object
        lastModified:
          type: string
          format: date-time
        priority:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
    OptimizationConfig:
      type: object
      properties:
        targetModel:
          type: string
        maxTokens:
          type: integer
        chunkingStrategy:
          type: string
          enum: [semantic, fixed, adaptive]
          default: semantic
        priority:
          type: string
          enum: [speed, quality, balanced]
          default: balanced
    ContextAnalysisRequest:
      type: object
      required: [projectId, documents]
      properties:
        projectId:
          type: string
          format: uuid
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        analysisType:
          type: string
          enum: [full, incremental, targeted]
          default: full
        optimization:
          $ref: '#/components/schemas/OptimizationConfig'
    ContextAnalysisResponse:
      type: object
      required: [contextId, projectId, status, documentsCount, estimatedChunks, estimatedCompletion, processingStage, createdAt]
      properties:
        contextId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        documentsCount:
          type: integer
        estimatedChunks:
          type: integer
        estimatedCompletion:
          type: string
          format: date-time
        processingStage:
          type: string
          enum: [queued, tokenizing, chunking, analyzing, optimizing, completed]
        createdAt:
          type: string
          format: date-time
    ContextOptimizationRequest:
      type: object
      required: [targetModel, maxTokens]
      properties:
        targetModel:
          type: string
        maxTokens:
          type: integer
          minimum: 100
          maximum: 200000
        useCase:
          type: string
          enum: [code_generation, documentation, debugging, analysis, chat]
        priority:
          type: string
          enum: [speed, quality, balanced]
          default: balanced
        preserveStructure:
          type: boolean
          default: true
        compressionLevel:
          type: string
          enum: [none, light, medium, aggressive]
          default: medium
    OptimizationMetrics:
      type: object
      required: [originalTokens, optimizedTokens, compressionRatio, chunksReduced, relevancePreserved]
      properties:
        originalTokens:
          type: integer
        optimizedTokens:
          type: integer
        compressionRatio:
          type: number
        chunksReduced:
          type: integer
        relevancePreserved:
          type: number
          minimum: 0
          maximum: 1
    ContextOptimizationResponse:
      type: object
      required: [contextId, optimizedAt, optimization, metadata]
      properties:
        contextId:
          type: string
          format: uuid
        optimizedAt:
          type: string
          format: date-time
        optimization:
          $ref: '#/components/schemas/OptimizationMetrics'
        metadata:
          type: object
          required: [targetModel, processingTime]
          properties:
            targetModel:
              type: string
            useCase:
              type: string
            processingTime:
              type: number
    ContextChunk:
      type: object
      required: [id, content, tokens, relevanceScore, priority, documentId, documentTitle, documentType, chunkIndex, metadata]
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        tokens:
          type: integer
        relevanceScore:
          type: number
          minimum: 0
          maximum: 1
        priority:
          type: number
          minimum: 0
          maximum: 1
        documentId:
          type: string
        documentTitle:
          type: string
        documentType:
          type: string
        chunkIndex:
          type: integer
        metadata:
          type: object
          required: [keywords, entities, concepts, complexity]
          properties:
            keywords:
              type: array
              items:
                type: string
            entities:
              type: array
              items:
                type: string
            concepts:
              type: array
              items:
                type: string
            complexity:
              type: number
              minimum: 0
              maximum: 1
    ContextChunksResponse:
      type: object
      required: [contextId, chunks, totalTokens, totalChunks, generatedAt]
      properties:
        contextId:
          type: string
          format: uuid
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ContextChunk'
        totalTokens:
          type: integer
        totalChunks:
          type: integer
        generatedAt:
          type: string
          format: date-time

    # Embedding schemas
...
    ErrorResponse:
      type: object
      required: [error, message, timestamp]
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time